test(taos): modify sql and stmt2 bench
