+++ taos/Cargo.toml
@@ -41,6 +41,7 @@ serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 tokio = { version = "1", features = ["rt-multi-thread"] }
 rand = "0.8.5"
+flume = "0.11"
 
 [[bench]]
 harness = false
@@ -60,6 +61,14 @@ ws = ["taos-ws"]
 ws-native-tls = ["ws", "taos-ws/native-tls"]
 ws-native-tls-vendored = ["ws", "taos-ws/native-tls-vendored"]
 ws-rustls = ["ws", "taos-ws/rustls"]
-ws-rustls-ring-crypto-provider = ["ws", "taos-ws/rustls", "taos-ws/rustls-ring-crypto-provider"]
-ws-rustls-aws-lc-crypto-provider = ["ws", "taos-ws/rustls", "taos-ws/rustls-aws-lc-crypto-provider"]
+ws-rustls-ring-crypto-provider = [
+    "ws",
+    "taos-ws/rustls",
+    "taos-ws/rustls-ring-crypto-provider",
+]
+ws-rustls-aws-lc-crypto-provider = [
+    "ws",
+    "taos-ws/rustls",
+    "taos-ws/rustls-aws-lc-crypto-provider",
+]
 deflate = ["ws", "taos-ws/deflate"]
